Rework version script
