+++ Jenkinsfile
@@ -33,40 +33,40 @@ node('linux && docker') {
         junit 'packages/*/reports/*.xml'
       }
 
-      // stage('Cypress Test') {
-      //   sh 'apt-get -y install libgtk2.0-0 libgtk-3-0 libnotify-dev libgconf-2-4 libgbm-dev libnss3 libasound2 xauth xvfb'
-      //   sh 'rm -rf /var/lib/apt/lists/*'
-      //   sh 'cd packages/atomic && npx cypress install'
-      //   sh 'npm start & npx wait-on http://localhost:3333'
-      //   sh 'NO_COLOR=1 npm run cypress:test'
-      // }
+       stage('Cypress Test') {
+         sh 'apt-get -y install libgtk2.0-0 libgtk-3-0 libnotify-dev libgconf-2-4 libgbm-dev libnss3 libasound2 xauth xvfb'
+         sh 'rm -rf /var/lib/apt/lists/*'
+         sh 'cd packages/atomic && npx cypress install'
+         sh 'npm start & npx wait-on http://localhost:3333'
+         sh 'NO_COLOR=1 npm run cypress:test'
+       }
     }
 
-    //if (!isMaster) {
-    //  return
-    //}
+    if (!isMaster) {
+      return
+    }
 
     stage('Clean working directory') {
       sh 'git checkout -- .'
       sh 'git clean -f'
     }
 
-    //withDockerContainer(image: 'node:14', args: '-u=root') {
-    //  stage('Bump version') {
-    //    withCredentials([
-    //    usernameColonPassword(credentialsId: 'github-commit-token', variable: 'GH_CREDENTIALS')]) {
-    //      sh 'npm run bump:version'
-    //    }
-    //  }
+    withDockerContainer(image: 'node:14', args: '-u=root') {
+      stage('Bump version') {
+        withCredentials([
+        usernameColonPassword(credentialsId: 'github-commit-token', variable: 'GH_CREDENTIALS')]) {
+          sh 'npm run bump:version:pre'
+        }
+     }
 
-    //  stage('Npm publish') {
-    //    withCredentials([
-    //    string(credentialsId: 'NPM_TOKEN', variable: 'NPM_TOKEN')]) {
-    //      sh "echo //registry.npmjs.org/:_authToken=${NPM_TOKEN} > ~/.npmrc"
-    //      sh 'npm run npm:publish:alpha || true'
-    //    }
-    //  }
-    //}
+      stage('Npm publish') {
+        withCredentials([
+        string(credentialsId: 'NPM_TOKEN', variable: 'NPM_TOKEN')]) {
+          sh "echo //registry.npmjs.org/:_authToken=${NPM_TOKEN} > ~/.npmrc"
+          sh 'npm run npm:publish:alpha || true'
+        }
+      }
+    }
 
     withDockerContainer(image: '458176070654.dkr.ecr.us-east-1.amazonaws.com/jenkins/deployment_package:v7') {
       stage('Veracode package') {
