+++ JenkinsfileProductionRelease
@@ -2,15 +2,20 @@ node('linux && docker') {
   checkout scm
 
   withDockerContainer(image: 'node:14', args: '-u=root') {
+
     stage('Graduate') {
-      print "GRADUATE SCRIPT VERSION = ${params.VERSION} HASH = ${params.COMMIT_HASH}"
+      withCredentials([
+      usernameColonPassword(credentialsId: 'github-commit-token', variable: 'GH_CREDENTIALS')]) {
+          sh 'npm run bump:version:graduate'
+      }
     }
 
     stage('Npm publish') {
-      withCredentials([
-      string(credentialsId: 'NPM_TOKEN', variable: 'NPM_TOKEN')]) {
-        print "PUBLISH"
+        withCredentials([
+        string(credentialsId: 'NPM_TOKEN', variable: 'NPM_TOKEN')]) {
+          sh "echo //registry.npmjs.org/:_authToken=${NPM_TOKEN} > ~/.npmrc"
+          sh 'npm run npm:publish'
+        }
       }
-    }
   }
 }
