+++ package.json
@@ -14,7 +14,8 @@
     "report:bundle-size:time-series": "node ./scripts/reports/bundle-size/time-series.js",
     "cypress:test": "lerna run cypress:test",
     "commit": "git-cz",
-    "bump:version": "node ./scripts/deploy/bump-version.js",
+    "bump:version:pre": "node ./scripts/deploy/bump-version.js",
+    "bump:version:graduate": "node ./scripts/deploy/bump-version.js --graduate",
     "version": "lerna version --conventional-commits",
     "version:pre": "lerna version --conventional-commits --conventional-prerelease",
     "version:graduate": "lerna version --conventional-commits --conventional-graduate",
